pyproject.toml
[tool.poetry]
name = "py-web-13"
version = "0.1.0"
description = ""
authors = ["checkmateLL <saoirse.girl47@gamil.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"
alembic = "^1.14.0"
email-validator = "^2.2.0"
cloudinary = "^1.41.0"
pydantic-settings = "^2.6.1"
redis = {extras = ["asyncio"], version = "^5.2.0"}
fastapi-mail = "^1.4.2"
python-jose = "^3.3.0"
python-multipart = "^0.0.17"
slowapi = "^0.1.9"
bcrypt = "4.0.1"
passlib = "1.7.4"
sphinx = "^8.1.3"
pytest = "^8.3.4"
httpx = "^0.28.1"
pytest-asyncio = "^0.25.0"
jwt = "^1.3.1"
aiosqlite = "^0.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]